# headless/api/documentation/openapi.yaml
openapi: "3.0.0"
info:
  title: "phpBB 4 Headless API"
  version: "1.0.0"
  description: |
    RESTful API layer for phpBB 4.
    Provides all forum functionality via JSON for modern frontend applications.
  contact:
    name: "EBTURK"
servers:
  - url: "/api/v1"
    description: "API v1"
security:
  - BearerAuth: []
paths:
...
  /topics/{topic_id}/posts:
    get:
      tags: [Posts]
      summary: "List posts"
      operationId: "postsList"
      parameters:
        - name: topic_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "Post list"
    post:
      tags: [Posts]
      summary: "Create post"
      operationId: "postCreate"
      parameters:
        - name: topic_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: "Post created"
  /posts/{post_id}:
    put:
      tags: [Posts]
      summary: "Update post"
      operationId: "postUpdate"
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "Post updated"
    delete:
      tags: [Posts]
      summary: "Delete post"
      operationId: "postDelete"
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "Post deleted"
  /posts/{post_id}/report:
    post:
      tags: [Posts]
      summary: "Report post"
      operationId: "postReport"
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "Report submitted"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT